+++ CHANGELOG.md
@@ -8,6 +8,8 @@ All notable changes to this project will be documented in this file. This projec
 
 ### Fixed
 
+- [#38](https://github.com/luyadev/luya-module-admin/issues/38) Fixed ActiveWindow render composition keys problem with PHP 7.2.
+
 ### Changed
 
 ## 1.0.1 (5. January 2018)